requirements.txt
astroid==3.0.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.7
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
platformdirs==4.1.0
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
yapf==0.40.2
zipp==3.17.0
